+++ kaizo/utils/__init__.py
@@ -1,3 +1,4 @@
+from .common import split_by_first_dot
 from .entry import DictEntry, Entry, FieldEntry, ListEntry, ModuleEntry
 from .fn import FnWithKwargs
 
@@ -8,4 +9,5 @@ __all__ = (
     "FnWithKwargs",
     "ListEntry",
     "ModuleEntry",
+    "split_by_first_dot",
 )