+++ tests/test_parser_errors.py
@@ -4,6 +4,10 @@ import pytest
 
 from kaizo import ConfigParser
 
+dummy_config = """
+dummy: this_is_dummy
+"""
+
 missing_config = """
 local: missing.py
 """
@@ -12,6 +16,31 @@ not_exist_config = """
 bad:
   module: does_not_exist
   source: thing
+"""
+
+invalid_import_type = """
+import: not_a_dict
+"""
+
+missing_import_file = """
+import:
+  miss: missing_module.yml
+"""
+
+module_not_given_config = """
+bad_ref: missing_module.some_key
+"""
+
+keyword_not_found_config = """
+import:
+  dummy: {tmp_path}/dummy.yml
+bad_ref: missing_module.some_key
+"""
+
+entry_not_found_config = """
+import:
+  dummy: {tmp_path}/dummy.yml
+bad_key: dummy.not_here
 """
 
 
@@ -30,4 +59,56 @@ def test_import_error(tmp_path: Path) -> None:
     parser = ConfigParser(cfg)
 
     with pytest.raises(ImportError):
+        parser.parse()
+
+
+def test_invalid_import_type(tmp_path: Path) -> None:
+    cfg = tmp_path / "cfg.yml"
+    cfg.write_text(invalid_import_type)
+
+    with pytest.raises(TypeError):
+        ConfigParser(cfg)
+
+
+def test_missing_import_file(tmp_path: Path) -> None:
+    cfg = tmp_path / "cfg.yml"
+    cfg.write_text(missing_import_file)
+
+    with pytest.raises(FileNotFoundError):
+        ConfigParser(cfg)
+
+
+def test_module_not_given(tmp_path: Path) -> None:
+    cfg = tmp_path / "cfg.yml"
+    cfg.write_text(module_not_given_config)
+
+    parser = ConfigParser(cfg)
+
+    with pytest.raises(ValueError, match="import module is not given"):
+        parser.parse()
+
+
+def test_keyword_not_found(tmp_path: Path) -> None:
+    dummy_module = tmp_path / "dummy.yml"
+    dummy_module.write_text(dummy_config)
+
+    cfg = tmp_path / "cfg.yml"
+    cfg.write_text(keyword_not_found_config.format(tmp_path=tmp_path))
+
+    parser = ConfigParser(cfg)
+
+    with pytest.raises(ValueError, match="keyword not found, got missing_module"):
+        parser.parse()
+
+
+def test_entry_not_found(tmp_path: Path) -> None:
+    dummy_module = tmp_path / "dummy.yml"
+    dummy_module.write_text(dummy_config)
+
+    cfg = tmp_path / "cfg.yml"
+    cfg.write_text(entry_not_found_config.format(tmp_path=tmp_path))
+
+    parser = ConfigParser(cfg)
+
+    with pytest.raises(KeyError, match="entry not found, got not_here"):
         parser.parse()